Prepare workspace for ch story 434
