+++ src/components/Chart/PieChart.jsx
@@ -1,0 +1,4 @@
+import { H1 } from '@blueprintjs/core'
+import React from 'react'
+
+export const PieChart = () => <H1>Chart!</H1>
