+++ src/components/Layout/TableSection.jsx
@@ -3,10 +3,9 @@ import { styled } from 'linaria/react'
 import React from 'react'
 
 import { humanMine25 as rows } from '../../stubs/humanMine25'
-import { Table } from './Table'
-import { TableActionButtons, TablePagingButtons } from './TableButtons'
-
-const mineUrl = 'https://www.humanmine.org/humanmine'
+import { mineUrl } from '../../stubs/utils'
+import { Table } from '../Table/Table'
+import { TableActionButtons, TablePagingButtons } from '../Table/TableButtons'
 
 const TableCard = styled(Card)`
 	margin-bottom: 20px;
