+++ src/components/Layout/TableSection.story.jsx
@@ -4,7 +4,7 @@ import { TableSection } from './TableSection'
 
 export default {
 	component: TableSection,
-	title: 'Components/TableSection',
+	title: 'Components/Table and Chart',
 }
 
 export const Table = () => <TableSection />
